requirements.txt
Flask
flask-jwt-extended
mysql-connector-python
python-dotenv
flask-cors
